Cjyc dev



查看合并请求 !4

+++ cjyc-web-api/src/main/java/com/cjyc/web/api/controller/StoreController.java
@@ -1,16 +1,22 @@
 package com.cjyc.web.api.controller;
 
+import com.cjkj.common.model.ResultData;
+import com.cjkj.common.model.ReturnMsg;
 import com.cjyc.common.model.dto.web.store.StoreAddDto;
 import com.cjyc.common.model.dto.web.store.StoreQueryDto;
 import com.cjyc.common.model.dto.web.store.StoreUpdateDto;
+import com.cjyc.common.model.entity.Admin;
 import com.cjyc.common.model.entity.Store;
 import com.cjyc.common.model.enums.ResultEnum;
 import com.cjyc.common.model.util.BaseResultUtil;
 import com.cjyc.common.model.vo.ResultVo;
+import com.cjyc.web.api.feign.ISysDeptService;
 import com.cjyc.web.api.service.IStoreService;
 import com.github.pagehelper.PageInfo;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
+import io.swagger.annotations.ApiParam;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
@@ -18,6 +24,7 @@ import org.springframework.web.bind.annotation.*;
 import javax.annotation.Resource;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import java.sql.PreparedStatement;
 import java.util.List;
 
 /**
@@ -31,6 +38,8 @@ import java.util.List;
 public class StoreController {
     @Resource
     private IStoreService storeService;
+    @Autowired
+    private ISysDeptService sysDeptService;
 
     /**
      * 根据二级城市编码查询业务中心
@@ -59,6 +68,14 @@ public class StoreController {
     @ApiOperation(value = "删除", notes = "\t 请求接口为/remove/id格式")
     @PostMapping("/remove/{id}")
     public ResultVo remove(@PathVariable Long id) {
+        Store store = storeService.getById(id);
+        if (null == store) {
+            return BaseResultUtil.fail("业务中心不存在，请检查");
+        }
+        ResultData deleteRd = sysDeptService.delete(store.getDeptId());
+        if (!ReturnMsg.SUCCESS.getCode().equals(deleteRd.getCode())) {
+            return BaseResultUtil.fail("业务中心删除失败，原因：" + deleteRd.getMsg());
+        }
         boolean result = storeService.removeById(id);
         return result ? BaseResultUtil.success() : BaseResultUtil.fail(ResultEnum.FAIL.getMsg());
     }
@@ -83,4 +100,11 @@ public class StoreController {
         storeService.exportExcel(request,response);
     }
 
+    @ApiOperation(value = "业务中心下用户列表信息", notes = "查询用户关联此用户中心角色信息")
+    @GetMapping("/listAdminsByStoreId/{storeId}")
+    public ResultVo<List<Admin>> listAdminsByStoreId(
+            @ApiParam(name = "storeId", value = "业务中心标识", required = true)
+            @PathVariable Long storeId){
+        return storeService.listAdminsByStoreId(storeId);
+    }
 }
